web-animations/animation-timing/animation-timing-easing.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationTiming easing attribute test</title>
<meta name="assert" content="The timing function used to scale the time to produce easing effects">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationtiming-easing">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
function createAnimationsToTest(test, timing) {
    var animation1 = new Animation(createDiv(test), KEYFRAMES, timing);
    var animation2 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
    var animation3 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
    var animationGroup = new AnimationGroup([animation2], timing);
    var animationSequence = new AnimationSequence([animation3], timing);
    return [animation1, animation2, animation3];
}

function testLinearEasing(player, animation, node) {
    assert_approx_equals(window.getComputedStyle(animation.target).top,
        getExpectedTop(player.currentTime), COMPUTED_STYLE_EPSILON,
        'Wrong effect for node ' + type(node) + ' at time ' + player.currentTime);
    player.currentTime = ANIMATION_DURATION / 4;
    assert_approx_equals(window.getComputedStyle(animation.target).top,
        getExpectedTop(player.currentTime), COMPUTED_STYLE_EPSILON,
        'Wrong effect for node ' + type(node) + ' at time ' + player.currentTime);
    player.currentTime = ANIMATION_DURATION / 2;
    assert_approx_equals(window.getComputedStyle(animation.target).top,
        getExpectedTop(player.currentTime), COMPUTED_STYLE_EPSILON,
        'Wrong effect for node ' + type(node) + ' at time ' + player.currentTime);
    player.currentTime = ANIMATION_DURATION * 3 / 4;
    assert_approx_equals(window.getComputedStyle(animation.target).top,
        getExpectedTop(player.currentTime), COMPUTED_STYLE_EPSILON,
        'Wrong effect for node ' + type(node) + ' at time ' + player.currentTime);
    player.currentTime = ANIMATION_DURATION - 1;
    assert_approx_equals(window.getComputedStyle(animation.target).top,
        getExpectedTop(player.currentTime), COMPUTED_STYLE_EPSILON,
        'Wrong effect for node ' + type(node) + ' at time ' + player.currentTime);
}

function bezierCurve(p1, p2, t) {
    return 3 * t * ((1 - t) ^ 2) * p1 + 3 * (t ^ 2) * (1 - t) * p2 + t ^ 3;
}

async_test(function(t) {
    var timing = {
        duration: ANIMATION_DURATION
    };
    var animations = createAnimationsToTest(t, timing);
    var doneCounter = 0;
    animations.forEach(function(animation) {
        var node = animation;
        if (animation.parent !== null) {
            node = node.parent;
        }
        var player = document.timeline.play(node);
        assert_equals(node.computedTiming.easing, 'linear',
            'Wrong AnimationNode.computedTiming.easing value');

        player.ready.then(t.step_func(function() {
            testLinearEasing(player, animation, node);
            doneCounter++;
            if (doneCounter === animations.length) {
                t.done();
            }
        }), t.unreached_func('Current ready promise should not be rejected'));
    });
}, 'Test that default value of AnimationTiming.easing is \'linear\' ' +
    'and animations goes as linear');

async_test(function(t) {
    var timing = {
        duration: ANIMATION_DURATION,
        easing: 'linear'
    };
    var animations = createAnimationsToTest(t, timing);
    var doneCounter = 0;
    animations.forEach(function(animation) {
        var node = animation;
        if (animation.parent !== null) {
            node = node.parent;
        }
        var player = document.timeline.play(node);
        assert_equals(node.computedTiming.easing, 'linear',
            'Wrong AnimationNode.computedTiming.easing value');

        player.ready.then(t.step_func(function() {
            testLinearEasing(player, animation, node);
            doneCounter++;
            if (doneCounter === animations.length) {
                t.done();
            }
        }), t.unreached_func('Current ready promise should not be rejected'));
    });
}, 'Test that if value of AnimationTiming.easing is \'linear\' ' +
    'then animations goes as linear');

async_test(function(t) {
    var timing = {
        duration: ANIMATION_DURATION,
        easing: 'unsupported-value'
    };
    var animations = createAnimationsToTest(t, timing);
    var doneCounter = 0;
    animations.forEach(function(animation) {
        var node = animation;
        if (animation.parent !== null) {
            node = node.parent;
        }
        var player = document.timeline.play(node);
        assert_equals(node.computedTiming.easing, 'linear',
            'Wrong AnimationNode.computedTiming.easing value');

        player.ready.then(t.step_func(function() {
            testLinearEasing(player, animation, node);
            doneCounter++;
            if (doneCounter === animations.length) {
                t.done();
            }
        }), t.unreached_func('Current ready promise should not be rejected'));
    });
}, 'Test that unsupported value of AnimationTiming.easing is treated ' +
    'as \'linear\' and animations goes as linear');

</script>
</body>
